Improve model comparison app code
